Finish edit function for editing items
