+++ src/style.css
@@ -69,11 +69,13 @@ li {
 .checkbox {
   margin: 5px 0 0 0;
   flex: 2;
+  width: 10px;
 }
 
 .description {
   flex: 2;
   font-weight: bold;
+  margin: 5px;
 }
 
 .dots {
